+++ packages/components/src/components/tab-nav/tab-nav.tsx
@@ -35,8 +35,6 @@ export class TabNav {
   @Prop() small?: boolean = false;
   /** (optional) size  */
   @Prop() size: 'small' | 'large' = 'small';
-  /** (optional) autoFocus  */
-  @Prop() autoFocus: boolean = false;
   /** (optional) Injected CSS styles */
   @Prop() styles?: string;
 
@@ -151,7 +149,6 @@ export class TabNav {
     tabs.forEach((tab) => {
       const panel = tab.nextElementSibling;
       tab.setAttribute('aria-controls', panel.id);
-      tab.setAttribute('auto-focus', this.autoFocus.toString());
       panel.setAttribute('aria-labelledby', tab.id);
     });
     this.selectTab(selectedTab);
@@ -190,7 +187,7 @@ export class TabNav {
 
   render() {
     return (
-      <Host>
+      <Host class="scale-tab-nav">
         {this.styles && <style>{this.styles}</style>}
 
         <div part={this.getBasePartMap()} class={this.getCssClassMap()}>