+++ packages/storybook-vue/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.0.0-beta.131](https://github.com/telekom/scale/compare/v3.0.0-beta.130...v3.0.0-beta.131) (2023-04-18)
+
+
+### Features
+
+* add placeholder prop in date picker ([#1720](https://github.com/telekom/scale/issues/1720)) ([335c78f](https://github.com/telekom/scale/commit/335c78f244bd1061b25c46d4877ee75e43a34603))
+* adding close and closing event ([#1698](https://github.com/telekom/scale/issues/1698)) ([5c325d9](https://github.com/telekom/scale/commit/5c325d905e6b2d86e6431c22d7ba7cbb2c3e0364))
+
+
+
+
+
 # [3.0.0-beta.130](https://github.com/telekom/scale/compare/v3.0.0-beta.129...v3.0.0-beta.130) (2023-04-06)